Moved image processing-related functions into a folder called 'image_processing', and updated associated imports
