+++ src/cryoemservices/util/image_processing/shared.py
