+++ src/cryoemservices/wrappers/clem_align_and_merge.py
@@ -24,9 +24,11 @@ from defusedxml.ElementTree import parse
 from pydantic import BaseModel, Field, ValidationError, field_validator, model_validator
 from tifffile import TiffFile, imwrite
 
-from cryoemservices.util.align_images_using_mmi import align_images_using_mmi
-from cryoemservices.util.drift_correct_image import drift_correct_image
-from cryoemservices.util.image_processing import (
+from cryoemservices.util.image_processing.align_images_using_mmi import (
+    align_images_using_mmi,
+)
+from cryoemservices.util.image_processing.drift_correct_image import drift_correct_image
+from cryoemservices.util.image_processing.shared import (
     convert_to_rgb,
     flatten_image,
     is_grayscale_image,
